Refactor `java.sql` type related tests under `com.fasterxml.jackson.databind.ext`
